+++ app/controllers/carto/kuviz/visualizations_controller.rb
@@ -1,10 +1,15 @@
+require_dependency 'carto/helpers/frame_options_helper'
+
 module Carto
   module Kuviz
     class VisualizationsController < ApplicationController
+      include Carto::FrameOptionsHelper
 
       ssl_required
 
+      before_action :x_frame_options_allow, only: [:show, :show_protected]
       before_action :get_kuviz
+
       skip_before_filter :verify_authenticity_token, only: [:show_protected]
 
       def show