+++ services/user-mover/import_user.rb
@@ -510,6 +510,7 @@ module CartoDB
         Carto::User.find(user_id).api_keys.select(&:needs_setup?).each do |k|
           k.role_permission_queries.each { |q| superuser_user_pg_conn.query(q) }
           k.grant_ownership_role_privileges
+          k.save_cdb_conf_info
         end
       end